+++ include/ggi/ggi_ext.h
@@ -1,4 +1,4 @@
-/* $Id: ggi_ext.h,v 1.1 2001/05/12 23:03:18 cegger Exp $
+/* $Id: ggi_ext.h,v 1.2 2003/07/13 07:21:25 cegger Exp $
 ******************************************************************************
 
    Copyright (C) 1998      Andreas Beck		[becka@ggi-project.org]
@@ -41,7 +41,7 @@ typedef int ggi_extid;	/* Don't rely on that ! */
 
 __BEGIN_DECLS
 
-ggi_extid    ggiExtensionRegister(char *name, int size,
+ggi_extid    ggiExtensionRegister(char *name, size_t size,
 			int (*paramchange)(ggi_visual_t,int whatchanged));
 int          ggiExtensionUnregister(ggi_extid id);
 int	     ggiExtensionAttach(ggi_visual_t vis,ggi_extid id);
